Fix: Made structure testable. StringCatalogEnumLibrary module is introduced
